Use OGC_SERVER.LOCATION for internal communication

+++ geonode/layers/models.py
@@ -132,7 +132,7 @@ class Layer(ResourceBase):
         # with the WMS parser.
         p = "&".join("%s=%s"%item for item in params.items())
 
-        return ogc_server_settings.public_url + "wms/reflect?" + p
+        return ogc_server_settings.LOCATION + "wms/reflect?" + p
 
 
     def verify(self):
@@ -331,7 +331,7 @@ def geoserver_pre_save(instance, sender, **kwargs):
         * Metadata Links,
         * Point of Contact name and url
     """
-    url = ogc_server_settings.rest
+    url = ogc_server_settings.internal_rest
     try:
         gs_catalog = Catalog(url, _user, _password)
         gs_resource = gs_catalog.get_resource(instance.name)
@@ -410,7 +410,7 @@ def geoserver_post_save(instance, sender, **kwargs):
        The way keywords are implemented requires the layer
        to be saved to the database before accessing them.
     """
-    url = ogc_server_settings.rest
+    url = ogc_server_settings.internal_rest
     
     try:
         gs_catalog = Catalog(url, _user, _password)
@@ -565,7 +565,8 @@ def geoserver_post_save(instance, sender, **kwargs):
     #remove links that belong to and old address
 
     for link in instance.link_set.all():
-        if not urlparse(ogc_server_settings.public_url).hostname == urlparse(link.url).hostname:
+        if not urlparse(settings.SITEURL).hostname == urlparse(link.url).hostname and not \
+                    urlparse(ogc_server_settings.public_url).hostname == urlparse(link.url).hostname:
             link.delete()
 
     #Save layer attributes
@@ -657,7 +658,7 @@ def set_attributes(layer, overwrite=False):
 
     attribute_map = []
     if layer.storeType == "dataStore":
-        dft_url = ogc_server_settings.public_url + "wfs?" + urllib.urlencode({
+        dft_url = ogc_server_settings.LOCATION + "wfs?" + urllib.urlencode({
             "service": "wfs",
             "version": "1.0.0",
             "request": "DescribeFeatureType",
@@ -671,7 +672,7 @@ def set_attributes(layer, overwrite=False):
         except Exception:
             attribute_map = []
     elif layer.storeType == "coverageStore":
-        dc_url = ogc_server_settings.public_url + "wcs?" + urllib.urlencode({
+        dc_url = ogc_server_settings.LOCATION + "wcs?" + urllib.urlencode({
             "service": "wcs",
             "version": "1.1.0",
             "request": "DescribeCoverage",
